feat: add services, basic repositories. add all models and dtos. add yandex postman tests fix: fix bugs

+++ main-service/src/main/java/ru/practicum/ewm/controller/adm/UserAdminController.java
@@ -28,7 +28,7 @@ public class UserAdminController {
     }
 
     @DeleteMapping("/admin/users{userId}")
-    public void deleteUser(@PathVariable Long userId) {
+    public void deleteUser(@PathVariable Integer userId) {
         userAdminService.deleteUser(userId);
     }
